backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
dnspython==1.16.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
motor==2.4.0
numpy==2.2.6
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==3.13.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.47.3
tflite-runtime==2.14.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
